+++ detekt-api/api/detekt-api.api
@@ -304,24 +304,6 @@ public final class io/gitlab/arturbosch/detekt/api/Location$Companion {
 	public static synthetic fun startLineAndColumn$default (Lio/gitlab/arturbosch/detekt/api/Location$Companion;Lorg/jetbrains/kotlin/com/intellij/psi/PsiElement;IILjava/lang/Object;)Lorg/jetbrains/kotlin/diagnostics/PsiDiagnosticUtils$LineAndColumn;
 }
 
-public final class io/gitlab/arturbosch/detekt/api/Metric {
-	public fun <init> (DDI)V
-	public synthetic fun <init> (DDIILkotlin/jvm/internal/DefaultConstructorMarker;)V
-	public fun <init> (IIZI)V
-	public synthetic fun <init> (IIZIILkotlin/jvm/internal/DefaultConstructorMarker;)V
-	public final fun doubleThreshold ()D
-	public final fun doubleValue ()D
-	public final fun getConversionFactor ()I
-	public final fun getThreshold ()I
-	public final fun getValue ()I
-	public final fun isDouble ()Z
-	public fun toString ()Ljava/lang/String;
-}
-
-public final class io/gitlab/arturbosch/detekt/api/MetricKt {
-	public static final field DEFAULT_FLOAT_CONVERSION_FACTOR I
-}
-
 public abstract interface class io/gitlab/arturbosch/detekt/api/Notification {
 	public abstract fun getLevel ()Lio/gitlab/arturbosch/detekt/api/Notification$Level;
 	public abstract fun getMessage ()Ljava/lang/String;
@@ -459,14 +441,6 @@ public final class io/gitlab/arturbosch/detekt/api/TextLocation {
 	public final fun getStart ()I
 	public fun hashCode ()I
 	public fun toString ()Ljava/lang/String;
-}
-
-public class io/gitlab/arturbosch/detekt/api/ThresholdedCodeSmell : io/gitlab/arturbosch/detekt/api/CodeSmell {
-	public fun <init> (Lio/gitlab/arturbosch/detekt/api/Issue;Lio/gitlab/arturbosch/detekt/api/Entity;Lio/gitlab/arturbosch/detekt/api/Metric;Ljava/lang/String;Ljava/util/List;)V
-	public synthetic fun <init> (Lio/gitlab/arturbosch/detekt/api/Issue;Lio/gitlab/arturbosch/detekt/api/Entity;Lio/gitlab/arturbosch/detekt/api/Metric;Ljava/lang/String;Ljava/util/List;ILkotlin/jvm/internal/DefaultConstructorMarker;)V
-	public fun compact ()Ljava/lang/String;
-	public final fun getMetric ()Lio/gitlab/arturbosch/detekt/api/Metric;
-	public fun messageOrDescription ()Ljava/lang/String;
 }
 
 public abstract interface annotation class io/gitlab/arturbosch/detekt/api/UnstableApi : java/lang/annotation/Annotation {
