+++ detekt-rules-complexity/src/test/kotlin/io/gitlab/arturbosch/detekt/api/MetricSpec.kt
