+++ detekt-sample-extensions/src/main/kotlin/io/gitlab/arturbosch/detekt/sample/extensions/rules/TooManyFunctionsTwo.kt
@@ -1,12 +1,11 @@
 package io.gitlab.arturbosch.detekt.sample.extensions.rules
 
+import io.gitlab.arturbosch.detekt.api.CodeSmell
 import io.gitlab.arturbosch.detekt.api.Config
 import io.gitlab.arturbosch.detekt.api.Debt
 import io.gitlab.arturbosch.detekt.api.Entity
 import io.gitlab.arturbosch.detekt.api.Issue
-import io.gitlab.arturbosch.detekt.api.Metric
 import io.gitlab.arturbosch.detekt.api.Rule
-import io.gitlab.arturbosch.detekt.api.ThresholdedCodeSmell
 import io.gitlab.arturbosch.detekt.api.config
 import org.jetbrains.kotlin.psi.KtFile
 import org.jetbrains.kotlin.psi.KtNamedFunction
@@ -14,8 +13,6 @@ import org.jetbrains.kotlin.psi.KtNamedFunction
 /**
  * This rule is a copy of [TooManyFunctions] which allows the threshold to be configured
  * in the detekt configuration file.
- * A [ThresholdedCodeSmell] can then be used to provide more information about the
- * raised metric.
  */
 class TooManyFunctionsTwo(config: Config) : Rule(config) {
 
@@ -33,10 +30,9 @@ class TooManyFunctionsTwo(config: Config) : Rule(config) {
         super.visitKtFile(file)
         if (amount > allowedFunctions) {
             report(
-                ThresholdedCodeSmell(
+                CodeSmell(
                     issue,
                     entity = Entity.from(file),
-                    metric = Metric(value = amount, threshold = allowedFunctions),
                     message = "The file ${file.name} has $amount function declarations. " +
                         "The maximum number of allowed functions is specified with $allowedFunctions.",
                     references = emptyList()
